Move all OCR rendering methods to render.py, improve function naming
