+++ .gitignore
@@ -74,3 +74,5 @@ tags
 # VSCode
 .devcontainer
 devcontainer.json
+
+static/mmif-viz-cache
