+++ ocr.py
@@ -185,43 +185,6 @@ def paginate(frames_list):
 
     return {i: page for (i, page) in enumerate(pages)}
 
-
-def render_ocr_page(mmif_id, vid_path, view_id, page_number):
-    """
-    Iterate through frames and display the contents/alignments.
-    """
-    # Path for storing temporary images generated by cv2
-    cv2_vid = cv2.VideoCapture(vid_path)
-    tn_data_fname = cache.get_cache_root() / mmif_id / f"{view_id}-pages.json"
-    thumbnail_pages = json.load(open(tn_data_fname))
-    page = thumbnail_pages[str(page_number)]
-    prev_frame_cap = None
-    path = make_image_directory(mmif_id)
-    for frame_num, frame in page:
-        # If index is range instead of frame...
-        if frame.get("range"):
-            frame_num = (int(frame["range"][0]) + int(frame["range"][1])) / 2
-        cv2_vid.set(1, frame_num)
-        _, frame_cap = cv2_vid.read()
-        if frame_cap is None:
-            raise FileNotFoundError(f"Video file {vid_path} not found!")
-
-        # Double check histogram similarity of "repeat" frames -- if they're significantly different, un-mark as repeat
-        if prev_frame_cap is not None and frame["repeat"] and not is_duplicate_image(prev_frame_cap, frame_cap,
-                                                                                     cv2_vid):
-            frame["repeat"] = False
-        with tempfile.NamedTemporaryFile(dir=str(path), suffix=".jpg", delete=False) as tf:
-            cv2.imwrite(tf.name, frame_cap)
-            # "id" is just the name of the temp image file
-            frame["id"] = pathlib.Path(tf.name).name
-        prev_frame_cap = frame_cap
-
-    tn_page_html = render_template(
-        'ocr.html', vid_path=vid_path, view_id=view_id, page=page,
-        n_pages=len(thumbnail_pages), page_number=str(page_number), mmif_id=mmif_id)
-    return tn_page_html
-
-
 def make_image_directory(mmif_id):
     # Make path for temp OCR image files or clear image files if it exists
     path = cache.get_cache_root() / mmif_id / "img"
