+++ render.py
@@ -11,7 +11,10 @@ import displacy
 import traceback
 
 from helpers import *
-from ocr import get_ocr_frames, paginate, find_duplicates, save_json, render_ocr_page
+from ocr import get_ocr_frames, paginate, find_duplicates, save_json, make_image_directory, is_duplicate_image
+import cv2
+import json
+import tempfile
 
 import cache
 
@@ -113,7 +116,7 @@ def render_annotations(mmif, viz_id):
             elif abstract_view_type == "ASR":
                 tabs.append({"id": view.id, "tab_name": f"{app_shortname}-{view.id}", "html": render_asr_vtt(view, viz_id)})
             elif abstract_view_type == "OCR":
-                tabs.append({"id": view.id, "tab_name": f"{app_shortname}-{view.id}", "html": render_ocr(mmif, view, viz_id)})
+                tabs.append({"id": view.id, "tab_name": f"{app_shortname}-{view.id}", "html": prepare_and_render_ocr(mmif, view, viz_id)})
         
         except Exception as e:
             tabs.append({"id": view.id, 
@@ -178,7 +181,11 @@ def render_ner(mmif, view):
     ner_document = metadata.get('document')
     return displacy.visualize_ner(mmif, view, ner_document, current_app.root_path)
 
-def render_ocr(mmif, view, viz_id):
+def prepare_and_render_ocr(mmif, view, viz_id):
+    """
+    Prepares list of frames that will be passed back and forth between server
+    and client, and renders the first page of the OCR.
+    """
     vid_path = mmif.get_documents_by_type(DocumentTypes.VideoDocument)[0].location_path()
 
     ocr_frames = get_ocr_frames(view, mmif)
@@ -190,3 +197,39 @@ def render_ocr(mmif, view, viz_id):
     # Save page list as temp file
     save_json(frames_pages, view.id, viz_id)
     return render_ocr_page(viz_id, vid_path, view.id, 0)
+
+def render_ocr_page(mmif_id, vid_path, view_id, page_number):
+    """
+    Renders a single OCR page by iterating through frames and displaying the 
+    contents/alignments.
+    """
+    # Path for storing temporary images generated by cv2
+    cv2_vid = cv2.VideoCapture(vid_path)
+    tn_data_fname = cache.get_cache_root() / mmif_id / f"{view_id}-pages.json"
+    thumbnail_pages = json.load(open(tn_data_fname))
+    page = thumbnail_pages[str(page_number)]
+    prev_frame_cap = None
+    path = make_image_directory(mmif_id)
+    for frame_num, frame in page:
+        # If index is range instead of frame...
+        if frame.get("range"):
+            frame_num = (int(frame["range"][0]) + int(frame["range"][1])) / 2
+        cv2_vid.set(1, frame_num)
+        _, frame_cap = cv2_vid.read()
+        if frame_cap is None:
+            raise FileNotFoundError(f"Video file {vid_path} not found!")
+
+        # Double check histogram similarity of "repeat" frames -- if they're significantly different, un-mark as repeat
+        if prev_frame_cap is not None and frame["repeat"] and not is_duplicate_image(prev_frame_cap, frame_cap,
+                                                                                     cv2_vid):
+            frame["repeat"] = False
+        with tempfile.NamedTemporaryFile(dir=str(path), suffix=".jpg", delete=False) as tf:
+            cv2.imwrite(tf.name, frame_cap)
+            # "id" is just the name of the temp image file
+            frame["id"] = pathlib.Path(tf.name).name
+        prev_frame_cap = frame_cap
+
+    tn_page_html = render_template(
+        'ocr.html', vid_path=vid_path, view_id=view_id, page=page,
+        n_pages=len(thumbnail_pages), page_number=str(page_number), mmif_id=mmif_id)
+    return tn_page_html
